requirements.txt
aiogram>=3.0.0
aiohttp>=3.8.0
httpx>=0.24.0
Pillow>=10.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
gtts>=2.4.0
